Move packets struct in login server into separate files for AC and CA packets.
