+++ src/login/lclif.c
@@ -24,6 +24,8 @@
 #include "login/ipban.h"
 #include "login/login.h"
 #include "login/loginlog.h"
+#include "login/packets_ac_struct.h"
+#include "login/packets_ca_struct.h"
 #include "common/HPM.h"
 #include "common/cbasetypes.h"
 #include "common/db.h"
